Add thread safety controls to tuple and circular buffer

+++ Template/deque.hpp
@@ -6,6 +6,8 @@
 #include "../Errno/errno.hpp"
 #include "../CPP_class/class_nullptr.hpp"
 #include "../Libft/libft.hpp"
+#include "../PThread/mutex.hpp"
+#include "../PThread/pthread.hpp"
 #include <cstddef>
 #include <utility>
 #include "move.hpp"
@@ -21,12 +23,17 @@ class ft_deque
             DequeNode* _next;
         };
 
-        DequeNode*   _front;
-        DequeNode*   _back;
-        size_t       _size;
-        mutable int  _error_code;
+        DequeNode*           _front;
+        DequeNode*           _back;
+        size_t               _size;
+        mutable int          _error_code;
+        mutable pt_mutex*    _mutex;
+        bool                 _thread_safe_enabled;
 
         void    set_error(int error) const;
+        int     lock_internal(bool *lock_acquired) const;
+        void    unlock_internal(bool lock_acquired) const;
+        void    teardown_thread_safety();
 
     public:
         ft_deque();
@@ -37,6 +44,12 @@ class ft_deque
 
         ft_deque(ft_deque&& other) noexcept;
         ft_deque& operator=(ft_deque&& other) noexcept;
+
+        int     enable_thread_safety();
+        void    disable_thread_safety();
+        bool    is_thread_safe() const;
+        int     lock(bool *lock_acquired) const;
+        void    unlock(bool lock_acquired) const;
 
         void push_front(const ElementType& value);
         void push_front(ElementType&& value);
@@ -61,7 +74,9 @@ class ft_deque
 
 template <typename ElementType>
 ft_deque<ElementType>::ft_deque()
-    : _front(ft_nullptr), _back(ft_nullptr), _size(0), _error_code(ER_SUCCESS)
+    : _front(ft_nullptr), _back(ft_nullptr), _size(0),
+      _error_code(ER_SUCCESS), _mutex(ft_nullptr),
+      _thread_safe_enabled(false)
 {
     this->set_error(ER_SUCCESS);
     return ;
@@ -71,17 +86,22 @@ template <typename ElementType>
 ft_deque<ElementType>::~ft_deque()
 {
     this->clear();
+    this->teardown_thread_safety();
     return ;
 }
 
 template <typename ElementType>
 ft_deque<ElementType>::ft_deque(ft_deque&& other) noexcept
-    : _front(other._front), _back(other._back), _size(other._size), _error_code(other._error_code)
+    : _front(other._front), _back(other._back), _size(other._size),
+      _error_code(other._error_code), _mutex(other._mutex),
+      _thread_safe_enabled(other._thread_safe_enabled)
 {
     other._front = ft_nullptr;
     other._back = ft_nullptr;
     other._size = 0;
     other._error_code = ER_SUCCESS;
+    other._mutex = ft_nullptr;
+    other._thread_safe_enabled = false;
     return ;
 }
 
@@ -91,16 +111,21 @@ ft_deque<ElementType>& ft_deque<ElementType>::operator=(ft_deque&& other) noexce
     if (this != &other)
     {
         this->clear();
+        this->teardown_thread_safety();
         this->_front = other._front;
         this->_back = other._back;
         this->_size = other._size;
         this->_error_code = other._error_code;
+        this->_mutex = other._mutex;
+        this->_thread_safe_enabled = other._thread_safe_enabled;
         other._front = ft_nullptr;
         other._back = ft_nullptr;
         other._size = 0;
         other._error_code = ER_SUCCESS;
-        this->set_error(ER_SUCCESS);
+        other._mutex = ft_nullptr;
+        other._thread_safe_enabled = false;
     }
+    this->set_error(ER_SUCCESS);
     return (*this);
 }
 
@@ -113,12 +138,104 @@ void ft_deque<ElementType>::set_error(int error) const
 }
 
 template <typename ElementType>
+int ft_deque<ElementType>::enable_thread_safety()
+{
+    void     *memory;
+    pt_mutex *mutex_pointer;
+
+    if (this->_thread_safe_enabled && this->_mutex != ft_nullptr)
+    {
+        this->set_error(ER_SUCCESS);
+        return (0);
+    }
+    memory = cma_malloc(sizeof(pt_mutex));
+    if (memory == ft_nullptr)
+    {
+        this->set_error(FT_ERR_NO_MEMORY);
+        return (-1);
+    }
+    mutex_pointer = new(memory) pt_mutex();
+    if (mutex_pointer->get_error() != ER_SUCCESS)
+    {
+        int mutex_error;
+
+        mutex_error = mutex_pointer->get_error();
+        mutex_pointer->~pt_mutex();
+        cma_free(memory);
+        this->set_error(mutex_error);
+        return (-1);
+    }
+    this->_mutex = mutex_pointer;
+    this->_thread_safe_enabled = true;
+    this->set_error(ER_SUCCESS);
+    return (0);
+}
+
+template <typename ElementType>
+void ft_deque<ElementType>::disable_thread_safety()
+{
+    this->teardown_thread_safety();
+    this->set_error(ER_SUCCESS);
+    return ;
+}
+
+template <typename ElementType>
+bool ft_deque<ElementType>::is_thread_safe() const
+{
+    bool enabled;
+
+    enabled = (this->_thread_safe_enabled && this->_mutex != ft_nullptr);
+    const_cast<ft_deque<ElementType> *>(this)->set_error(ER_SUCCESS);
+    return (enabled);
+}
+
+template <typename ElementType>
+int ft_deque<ElementType>::lock(bool *lock_acquired) const
+{
+    int result;
+
+    result = this->lock_internal(lock_acquired);
+    if (result != 0)
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ft_errno);
+    else
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ER_SUCCESS);
+    return (result);
+}
+
+template <typename ElementType>
+void ft_deque<ElementType>::unlock(bool lock_acquired) const
+{
+    int entry_errno;
+
+    entry_errno = ft_errno;
+    this->unlock_internal(lock_acquired);
+    if (this->_mutex != ft_nullptr && this->_mutex->get_error() != ER_SUCCESS)
+        const_cast<ft_deque<ElementType> *>(this)->set_error(this->_mutex->get_error());
+    else
+    {
+        ft_errno = entry_errno;
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ft_errno);
+    }
+    return ;
+}
+
+template <typename ElementType>
 void ft_deque<ElementType>::push_front(const ElementType& value)
 {
-    DequeNode* node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
+    DequeNode* node;
+    bool       lock_acquired;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return ;
+    }
+    node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
     if (node == ft_nullptr)
     {
         this->set_error(FT_ERR_NO_MEMORY);
+        this->unlock_internal(lock_acquired);
         return ;
     }
     construct_at(&node->_data, value);
@@ -129,18 +246,29 @@ void ft_deque<ElementType>::push_front(const ElementType& value)
     else
         this->_front->_prev = node;
     this->_front = node;
-    ++this->_size;
+    this->_size += 1;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return ;
 }
 
 template <typename ElementType>
 void ft_deque<ElementType>::push_front(ElementType&& value)
 {
-    DequeNode* node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
+    DequeNode* node;
+    bool       lock_acquired;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return ;
+    }
+    node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
     if (node == ft_nullptr)
     {
         this->set_error(FT_ERR_NO_MEMORY);
+        this->unlock_internal(lock_acquired);
         return ;
     }
     construct_at(&node->_data, ft_move(value));
@@ -151,18 +279,29 @@ void ft_deque<ElementType>::push_front(ElementType&& value)
     else
         this->_front->_prev = node;
     this->_front = node;
-    ++this->_size;
+    this->_size += 1;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return ;
 }
 
 template <typename ElementType>
 void ft_deque<ElementType>::push_back(const ElementType& value)
 {
-    DequeNode* node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
+    DequeNode* node;
+    bool       lock_acquired;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return ;
+    }
+    node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
     if (node == ft_nullptr)
     {
         this->set_error(FT_ERR_NO_MEMORY);
+        this->unlock_internal(lock_acquired);
         return ;
     }
     construct_at(&node->_data, value);
@@ -173,18 +312,29 @@ void ft_deque<ElementType>::push_back(const ElementType& value)
     else
         this->_back->_next = node;
     this->_back = node;
-    ++this->_size;
+    this->_size += 1;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return ;
 }
 
 template <typename ElementType>
 void ft_deque<ElementType>::push_back(ElementType&& value)
 {
-    DequeNode* node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
+    DequeNode* node;
+    bool       lock_acquired;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return ;
+    }
+    node = static_cast<DequeNode*>(cma_malloc(sizeof(DequeNode)));
     if (node == ft_nullptr)
     {
         this->set_error(FT_ERR_NO_MEMORY);
+        this->unlock_internal(lock_acquired);
         return ;
     }
     construct_at(&node->_data, ft_move(value));
@@ -195,52 +345,77 @@ void ft_deque<ElementType>::push_back(ElementType&& value)
     else
         this->_back->_next = node;
     this->_back = node;
-    ++this->_size;
+    this->_size += 1;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return ;
 }
 
 template <typename ElementType>
 ElementType ft_deque<ElementType>::pop_front()
 {
+    bool        lock_acquired;
+    DequeNode  *node;
+    ElementType value;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return (ElementType());
+    }
     if (this->_front == ft_nullptr)
     {
         this->set_error(FT_ERR_EMPTY);
+        this->unlock_internal(lock_acquired);
         return (ElementType());
     }
-    DequeNode* node = this->_front;
+    node = this->_front;
     this->_front = node->_next;
     if (this->_front == ft_nullptr)
         this->_back = ft_nullptr;
     else
         this->_front->_prev = ft_nullptr;
-    ElementType value = ft_move(node->_data);
+    value = ft_move(node->_data);
     destroy_at(&node->_data);
     cma_free(node);
-    --this->_size;
+    this->_size -= 1;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return (value);
 }
 
 template <typename ElementType>
 ElementType ft_deque<ElementType>::pop_back()
 {
+    bool        lock_acquired;
+    DequeNode  *node;
+    ElementType value;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return (ElementType());
+    }
     if (this->_back == ft_nullptr)
     {
         this->set_error(FT_ERR_EMPTY);
+        this->unlock_internal(lock_acquired);
         return (ElementType());
     }
-    DequeNode* node = this->_back;
+    node = this->_back;
     this->_back = node->_prev;
     if (this->_back == ft_nullptr)
         this->_front = ft_nullptr;
     else
         this->_back->_next = ft_nullptr;
-    ElementType value = ft_move(node->_data);
+    value = ft_move(node->_data);
     destroy_at(&node->_data);
     cma_free(node);
-    --this->_size;
+    this->_size -= 1;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return (value);
 }
 
@@ -248,72 +423,136 @@ template <typename ElementType>
 ElementType& ft_deque<ElementType>::front()
 {
     static ElementType error_element = ElementType();
+    bool               lock_acquired;
+    ElementType       *value_pointer;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return (error_element);
+    }
     if (this->_front == ft_nullptr)
     {
         this->set_error(FT_ERR_EMPTY);
+        this->unlock_internal(lock_acquired);
         return (error_element);
     }
-    ElementType& reference = this->_front->_data;
+    value_pointer = &this->_front->_data;
     this->set_error(ER_SUCCESS);
-    return (reference);
+    this->unlock_internal(lock_acquired);
+    return (*value_pointer);
 }
 
 template <typename ElementType>
 const ElementType& ft_deque<ElementType>::front() const
 {
     static ElementType error_element = ElementType();
-    if (this->_front == ft_nullptr)
+    bool               lock_acquired;
+    const ElementType *value_pointer;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
     {
-        this->set_error(FT_ERR_EMPTY);
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ft_errno);
         return (error_element);
     }
-    const ElementType& reference = this->_front->_data;
+    if (this->_front == ft_nullptr)
+    {
+        const_cast<ft_deque<ElementType> *>(this)->set_error(FT_ERR_EMPTY);
+        this->unlock_internal(lock_acquired);
+        return (error_element);
+    }
+    value_pointer = &this->_front->_data;
     this->set_error(ER_SUCCESS);
-    return (reference);
+    this->unlock_internal(lock_acquired);
+    return (*value_pointer);
 }
 
 template <typename ElementType>
 ElementType& ft_deque<ElementType>::back()
 {
     static ElementType error_element = ElementType();
+    bool               lock_acquired;
+    ElementType       *value_pointer;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return (error_element);
+    }
     if (this->_back == ft_nullptr)
     {
         this->set_error(FT_ERR_EMPTY);
+        this->unlock_internal(lock_acquired);
         return (error_element);
     }
-    ElementType& reference = this->_back->_data;
+    value_pointer = &this->_back->_data;
     this->set_error(ER_SUCCESS);
-    return (reference);
+    this->unlock_internal(lock_acquired);
+    return (*value_pointer);
 }
 
 template <typename ElementType>
 const ElementType& ft_deque<ElementType>::back() const
 {
     static ElementType error_element = ElementType();
-    if (this->_back == ft_nullptr)
+    bool               lock_acquired;
+    const ElementType *value_pointer;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
     {
-        this->set_error(FT_ERR_EMPTY);
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ft_errno);
         return (error_element);
     }
-    const ElementType& reference = this->_back->_data;
+    if (this->_back == ft_nullptr)
+    {
+        const_cast<ft_deque<ElementType> *>(this)->set_error(FT_ERR_EMPTY);
+        this->unlock_internal(lock_acquired);
+        return (error_element);
+    }
+    value_pointer = &this->_back->_data;
     this->set_error(ER_SUCCESS);
-    return (reference);
+    this->unlock_internal(lock_acquired);
+    return (*value_pointer);
 }
 
 template <typename ElementType>
 size_t ft_deque<ElementType>::size() const
 {
-    size_t current_size = this->_size;
+    bool   lock_acquired;
+    size_t current_size;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ft_errno);
+        return (0);
+    }
+    current_size = this->_size;
     const_cast<ft_deque<ElementType> *>(this)->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
     return (current_size);
 }
 
 template <typename ElementType>
 bool ft_deque<ElementType>::empty() const
 {
-    bool result = (this->_size == 0);
+    bool lock_acquired;
+    bool is_empty;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        const_cast<ft_deque<ElementType> *>(this)->set_error(ft_errno);
+        return (true);
+    }
+    is_empty = (this->_size == 0);
     const_cast<ft_deque<ElementType> *>(this)->set_error(ER_SUCCESS);
-    return (result);
+    this->unlock_internal(lock_acquired);
+    return (is_empty);
 }
 
 template <typename ElementType>
@@ -331,9 +570,18 @@ const char* ft_deque<ElementType>::get_error_str() const
 template <typename ElementType>
 void ft_deque<ElementType>::clear()
 {
+    bool       lock_acquired;
+    DequeNode *node;
+
+    lock_acquired = false;
+    if (this->lock_internal(&lock_acquired) != 0)
+    {
+        this->set_error(ft_errno);
+        return ;
+    }
     while (this->_front != ft_nullptr)
     {
-        DequeNode* node = this->_front;
+        node = this->_front;
         this->_front = this->_front->_next;
         destroy_at(&node->_data);
         cma_free(node);
@@ -341,6 +589,71 @@ void ft_deque<ElementType>::clear()
     this->_back = ft_nullptr;
     this->_size = 0;
     this->set_error(ER_SUCCESS);
+    this->unlock_internal(lock_acquired);
+    return ;
+}
+
+template <typename ElementType>
+int ft_deque<ElementType>::lock_internal(bool *lock_acquired) const
+{
+    if (lock_acquired)
+        *lock_acquired = false;
+    if (!this->_thread_safe_enabled || this->_mutex == ft_nullptr)
+    {
+        ft_errno = ER_SUCCESS;
+        return (0);
+    }
+    this->_mutex->lock(THREAD_ID);
+    if (this->_mutex->get_error() != ER_SUCCESS)
+    {
+        if (this->_mutex->get_error() == FT_ERR_MUTEX_ALREADY_LOCKED)
+        {
+            bool state_lock_acquired;
+
+            state_lock_acquired = false;
+            ft_errno = ER_SUCCESS;
+            if (this->_mutex->lock_state(&state_lock_acquired) == 0)
+                this->_mutex->unlock_state(state_lock_acquired);
+            ft_errno = ER_SUCCESS;
+            return (0);
+        }
+        ft_errno = this->_mutex->get_error();
+        return (-1);
+    }
+    if (lock_acquired)
+        *lock_acquired = true;
+    ft_errno = ER_SUCCESS;
+    return (0);
+}
+
+template <typename ElementType>
+void ft_deque<ElementType>::unlock_internal(bool lock_acquired) const
+{
+    int entry_errno;
+
+    if (!lock_acquired || this->_mutex == ft_nullptr)
+        return ;
+    entry_errno = ft_errno;
+    this->_mutex->unlock(THREAD_ID);
+    if (this->_mutex->get_error() != ER_SUCCESS)
+    {
+        ft_errno = this->_mutex->get_error();
+        return ;
+    }
+    ft_errno = entry_errno;
+    return ;
+}
+
+template <typename ElementType>
+void ft_deque<ElementType>::teardown_thread_safety()
+{
+    if (this->_mutex != ft_nullptr)
+    {
+        this->_mutex->~pt_mutex();
+        cma_free(this->_mutex);
+        this->_mutex = ft_nullptr;
+    }
+    this->_thread_safe_enabled = false;
     return ;
 }
